Fix how old services get disabled/restart
systemctl disable won't work here because loki-node@ is already a
symlink to oxen-node@, and this makes systemctl disable treat the
request as a request to disable `oxen-node@NN` -- but that doesn't work
because it isn't enabled under that name.

This is annoying.

Also don't stop old/start new but instead just restart new because systemd will
already see the old services as the new name because loki-node@ is
already a symlink to oxen-node@.